ENHANCEMENT Add PSR-4 autoloading and namespacing
