+++ _config.php
@@ -1,5 +1,5 @@
 <?php
 
-if(!defined('CRONTASK_MODULE_PATH')) {
-	define('CRONTASK_MODULE_PATH', dirname(__DIR__));
+if (!defined('CRONTASK_MODULE_PATH')) {
+    define('CRONTASK_MODULE_PATH', dirname(__DIR__));
 }
