+++ tests/CronTaskControllerTest.php
@@ -1,17 +1,32 @@
 <?php
 
+namespace SilverStripe\CronTask\Tests;
+
 use Cron\CronExpression;
 use SilverStripe\ORM\FieldType\DBDatetime;
+use SilverStripe\CronTask\Controllers\CronTaskController;
+use SilverStripe\CronTask\CronTaskStatus;
+use SilverStripe\CronTask\Controllers\CronTask;
+use SilverStripe\Dev\SapphireTest;
 
-class CronTaskControllerTest extends SilverStripe\Dev\SapphireTest
+/**
+ * @package crontask
+ */
+class CronTaskControllerTest extends SapphireTest
 {
-
+    /**
+     * {@inheritDoc}
+     * @var bool
+     */
     protected $usesDatabase = true;
 
+    /**
+     * {@inheritDoc}
+     */
     public function setUp()
     {
         parent::setUp();
-        CronTaskTest_TestCron::$times_run = 0;
+        CronTaskTest\TestCron::$times_run = 0;
     }
 
     /**
@@ -20,7 +35,7 @@ class CronTaskControllerTest extends SilverStripe\Dev\SapphireTest
     public function testIsTaskDue()
     {
         $runner = CronTaskController::create();
-        $task = new CronTaskTest_TestCron();
+        $task = new CronTaskTest\TestCron();
         $cron = CronExpression::factory($task->getSchedule());
 
         // Assuming first run, match the exact time (seconds are ignored)
@@ -35,7 +50,7 @@ class CronTaskControllerTest extends SilverStripe\Dev\SapphireTest
 
         // Mock a run and test that subsequent runs are properly scheduled
         DBDatetime::set_mock_now('2010-06-20 13:30:10');
-        CronTaskStatus::update_status('CronTaskTest_TestCron', true);
+        CronTaskStatus::update_status('SilverStripe\\CronTask\\Tests\\CronTaskTest\\TestCron', true);
 
         // Job prior to next hour mark should not run
         DBDatetime::set_mock_now('2010-06-20 13:40:00');
@@ -57,55 +72,37 @@ class CronTaskControllerTest extends SilverStripe\Dev\SapphireTest
     {
         $runner = CronTaskController::create();
         $runner->setQuiet(true);
-        $task = new CronTaskTest_TestCron();
+        $task = new CronTaskTest\TestCron();
 
         // Assuming first run, match the exact time (seconds are ignored)
-        $this->assertEquals(0, CronTaskTest_TestCron::$times_run);
+        $this->assertEquals(0, CronTaskTest\TestCron::$times_run);
         DBDatetime::set_mock_now('2010-06-20 13:00:10');
         $runner->runTask($task);
-        $this->assertEquals(1, CronTaskTest_TestCron::$times_run);
+        $this->assertEquals(1, CronTaskTest\TestCron::$times_run);
 
         // Test that re-requsting the task in the same minute do not retrigger another run
         DBDatetime::set_mock_now('2010-06-20 13:00:40');
         $runner->runTask($task);
-        $this->assertEquals(1, CronTaskTest_TestCron::$times_run);
+        $this->assertEquals(1, CronTaskTest\TestCron::$times_run);
 
         // Job prior to next hour mark should not run
         DBDatetime::set_mock_now('2010-06-20 13:40:00');
         $runner->runTask($task);
-        $this->assertEquals(1, CronTaskTest_TestCron::$times_run);
+        $this->assertEquals(1, CronTaskTest\TestCron::$times_run);
 
         // Jobs just after the next hour mark should run
         DBDatetime::set_mock_now('2010-06-20 14:10:00');
         $runner->runTask($task);
-        $this->assertEquals(2, CronTaskTest_TestCron::$times_run);
+        $this->assertEquals(2, CronTaskTest\TestCron::$times_run);
 
         // Jobs run on the exact next expected date should run
         DBDatetime::set_mock_now('2010-06-20 15:00:00');
         $runner->runTask($task);
-        $this->assertEquals(3, CronTaskTest_TestCron::$times_run);
+        $this->assertEquals(3, CronTaskTest\TestCron::$times_run);
 
         // Jobs somehow delayed a whole day should be run
         DBDatetime::set_mock_now('2010-06-21 13:40:00');
         $runner->runTask($task);
-        $this->assertEquals(4, CronTaskTest_TestCron::$times_run);
-    }
-}
-
-
-class CronTaskTest_TestCron implements SilverStripe\Dev\TestOnly, CronTask
-{
-
-    public static $times_run = 0;
-
-    public function getSchedule()
-    {
-        // Use hourly schedule
-        return '0 * * * *';
-    }
-
-    public function process()
-    {
-        ++self::$times_run;
+        $this->assertEquals(4, CronTaskTest\TestCron::$times_run);
     }
 }
